refactor(ema_db): Group continuous sensor repos
