+++ lib/src/core/ema_db/continuous_sensor_data/pedestrian_status/models/pedestrian_status.dart
