+++ lib/src/core/ema_db/continuous_sensor_data/pedestrian_status/pedestrian_status_repository.dart
@@ -1,24 +1,24 @@
 import 'package:mdigits/src/core/ema_db/datasources/remote_datasource.dart';
-import 'package:mdigits/src/core/ema_db/step_count/models/step_count_model.dart';
+import 'package:mdigits/src/core/ema_db/continuous_sensor_data/pedestrian_status/models/pedestrian_status_model.dart';
 
-/// Provides a simple interface for managing [StepCount] data.
-class StepCountRepository {
+/// Provides a simple interface for managing [PedestrianStatus] data.
+class PedestrianStatusRepository {
   final RemoteDataSource _remoteDataSource;
 
-  StepCountRepository({
+  PedestrianStatusRepository({
     required RemoteDataSource remoteDataSource,
   }) : _remoteDataSource = remoteDataSource;
 
-  /// Saves [stepCount] to the remote database.
+  /// Saves [pedestrianStatus] to the remote database.
   ///
   /// [pathRemoteDB] must be a valid path that can be used to create a collection
   /// reference in the Firestore database.
   Future<void> save({
-    required StepCountModel stepCount,
+    required PedestrianStatusModel pedestrianStatus,
     required String pathRemoteDB,
   }) async {
     await _remoteDataSource.saveEMAModel(
-      emaModel: stepCount,
+      emaModel: pedestrianStatus,
       path: pathRemoteDB,
     );
   }
