+++ lib/src/core/ema_db/continuous_sensor_data/step_count/step_count_repository.dart
@@ -1,24 +1,24 @@
 import 'package:mdigits/src/core/ema_db/datasources/remote_datasource.dart';
-import 'package:mdigits/src/core/ema_db/pedestrian_status/models/pedestrian_status_model.dart';
+import 'package:mdigits/src/core/ema_db/continuous_sensor_data/step_count/models/step_count_model.dart';
 
-/// Provides a simple interface for managing [PedestrianStatus] data.
-class PedestrianStatusRepository {
+/// Provides a simple interface for managing [StepCount] data.
+class StepCountRepository {
   final RemoteDataSource _remoteDataSource;
 
-  PedestrianStatusRepository({
+  StepCountRepository({
     required RemoteDataSource remoteDataSource,
   }) : _remoteDataSource = remoteDataSource;
 
-  /// Saves [pedestrianStatus] to the remote database.
+  /// Saves [stepCount] to the remote database.
   ///
   /// [pathRemoteDB] must be a valid path that can be used to create a collection
   /// reference in the Firestore database.
   Future<void> save({
-    required PedestrianStatusModel pedestrianStatus,
+    required StepCountModel stepCount,
     required String pathRemoteDB,
   }) async {
     await _remoteDataSource.saveEMAModel(
-      emaModel: pedestrianStatus,
+      emaModel: stepCount,
       path: pathRemoteDB,
     );
   }
