+++ test/src/core/ema_db/continuous_sensor_data/pedestrian_status/pedestrian_status_model_test_cases.dart
@@ -1,5 +1,5 @@
-import 'package:mdigits/src/core/ema_db/pedestrian_status/models/pedestrian_status_model.dart';
-import 'package:mdigits/src/core/ema_db/pedestrian_status/models/pedestrian_status.dart';
+import 'package:mdigits/src/core/ema_db/continuous_sensor_data/pedestrian_status/models/pedestrian_status_model.dart';
+import 'package:mdigits/src/core/ema_db/continuous_sensor_data/pedestrian_status/models/pedestrian_status.dart';
 
 final PedestrianStatusModel testPedestrianStatusModel = PedestrianStatusModel(
   participantId: 'testParticipant',
