+++ test/src/core/ema_db/continuous_sensor_data/step_count/step_count_model_test_cases.dart
@@ -1,4 +1,4 @@
-import 'package:mdigits/src/core/ema_db/step_count/models/step_count_model.dart';
+import 'package:mdigits/src/core/ema_db/continuous_sensor_data/step_count/models/step_count_model.dart';
 
 final StepCountModel testStepCountModel = StepCountModel(
   participantId: 'testParticipant',
